refactor: 重构组件结构
